src/Flask/requirements.txt
flask
flask-cors
requests
python-dotenv
google-cloud-storage